scripts/requirements.txt
feedparser==6.0.11
python-frontmatter==1.0.1
requests==2.32.3
python-dateutil==2.9.0.post0
